+++ BookGen.Gui/XmlEntities/XWindow.cs
@@ -18,6 +18,7 @@ namespace BookGen.Gui.XmlEntities
         [XmlArrayItem(nameof(XButton), typeof(XButton))]
         [XmlArrayItem(nameof(XTextBlock), typeof(XTextBlock))]
         [XmlArrayItem(nameof(XCheckBox), typeof(XCheckBox))]
+        [XmlArrayItem(nameof(XRadioGroup), typeof(XRadioGroup))]
         [XmlArrayItem(nameof(XListBox), typeof(XListBox))]
         [XmlArrayItem(nameof(XTextBox), typeof(XTextBox))]
         [XmlArrayItem(nameof(XSPlitView), typeof(XSPlitView))]