+++ Changelog.md
@@ -4,6 +4,8 @@
 * New: Launcher has more modern Message boxes
 * New: Added more options to the console Gui
 * New: Log output to JSON with the global -js argument.
+* Change: Launcher folder list is now independent from version
+* Fix: Launcher JumpList was not starting the shell in the folder
 
 
 # 2021.09.12