+++ Changelog.md
@@ -8,9 +8,10 @@
 * Change: Autocomplete handles global arguments.
 * Change: Launcher folder list is now independent from version
 * Change: Now requires .NET 6 runtime
+* Change: Configuration can be in yml format
+* Change: Init subcommand now allows to select yml as config format
 * Fix: Launcher JumpList was not starting the shell in the folder
 * Fix: Auto updater now works
-
 
 # 2021.09.12
 * New: Added an Update command to make future updates easy