+++ Changelog.md
@@ -5,6 +5,9 @@
 * New: Added more options to the console Gui
 * New: Log output to JSON with the global -js argument.
 * New: Assembly documenting to a single markdown file (usefull for wiki pages)
+* New: Installer based on Inno setup
+* Change: Updated bootstrap to 5.1
+* Change: BookGen shell now displays git info in prompt, if folder is git repo.
 * Change: Autocomplete handles global arguments.
 * Change: Launcher folder list is now independent from version
 * Change: Now requires .NET 6 runtime